refactor: migrate project form old Makefile to CMake build system. - Raylib is now being builded by CMake automatically, so the game can be   build on Windows/macOS/Linux without the need to download Raylib.
- Fixed 'segfault' in read_from_file().

- Fixed 'core bumped' in show_menu(), Double closeWindow(), one in
  show_menue() and the another in main().

- Used raylib's GetApplicationDirectory() instead of bare relative paths,
  so file loading doesn't depend on the current working directory.

- database.txt now opens with "a+" so it's created if missing,
  instead of crashing when it doesn't exist.
