--- database.txt
@@ -1,10 +1,0 @@
-3
-0
-0
-0
-3
-7
-0
-0
-0
-0
