--- main.dSYM/Contents/Resources/Relocations/aarch64/main.yml
@@ -1,24 +1,0 @@
----
-triple:          'arm64-apple-darwin'
-binary-path:     main
-relocations:
-  - { offsetInCU: 0x26, offset: 0x26, size: 0x8, addend: 0x0, symName: _read_from_file, symObjAddr: 0x0, symBinAddr: 0x1000023F4, symSize: 0x10C }
-  - { offsetInCU: 0x3F, offset: 0x3F, size: 0x8, addend: 0x0, symName: _game, symObjAddr: 0x48A0, symBinAddr: 0x100008000, symSize: 0x0 }
-  - { offsetInCU: 0x15D, offset: 0x15D, size: 0x8, addend: 0x0, symName: _block_stack, symBinAddr: 0x100008050, symSize: 0x0 }
-  - { offsetInCU: 0x1A9, offset: 0x1A9, size: 0x8, addend: 0x0, symName: _read_from_file, symObjAddr: 0x0, symBinAddr: 0x1000023F4, symSize: 0x10C }
-  - { offsetInCU: 0x1E9, offset: 0x1E9, size: 0x8, addend: 0x0, symName: _write_to_file, symObjAddr: 0x10C, symBinAddr: 0x100002500, symSize: 0x90 }
-  - { offsetInCU: 0x20D, offset: 0x20D, size: 0x8, addend: 0x0, symName: _init_entity, symObjAddr: 0x19C, symBinAddr: 0x100002590, symSize: 0x4C }
-  - { offsetInCU: 0x26D, offset: 0x26D, size: 0x8, addend: 0x0, symName: _check_collision, symObjAddr: 0x1E8, symBinAddr: 0x1000025DC, symSize: 0xF4 }
-  - { offsetInCU: 0x2A3, offset: 0x2A3, size: 0x8, addend: 0x0, symName: _reset_blocks, symObjAddr: 0x2DC, symBinAddr: 0x1000026D0, symSize: 0xF8 }
-  - { offsetInCU: 0x31B, offset: 0x31B, size: 0x8, addend: 0x0, symName: _reset_ball, symObjAddr: 0x3D4, symBinAddr: 0x1000027C8, symSize: 0x34 }
-  - { offsetInCU: 0x34D, offset: 0x34D, size: 0x8, addend: 0x0, symName: _random_x, symObjAddr: 0x408, symBinAddr: 0x1000027FC, symSize: 0x40 }
-  - { offsetInCU: 0x366, offset: 0x366, size: 0x8, addend: 0x0, symName: _init_game, symObjAddr: 0x448, symBinAddr: 0x10000283C, symSize: 0x98 }
-  - { offsetInCU: 0x3A6, offset: 0x3A6, size: 0x8, addend: 0x0, symName: _MENU_TITLE, symObjAddr: 0x4E0, symBinAddr: 0x1000028D4, symSize: 0xB4 }
-  - { offsetInCU: 0x3E6, offset: 0x3E6, size: 0x8, addend: 0x0, symName: _BUILD_BUTTON, symObjAddr: 0x594, symBinAddr: 0x100002988, symSize: 0xB4 }
-  - { offsetInCU: 0x426, offset: 0x426, size: 0x8, addend: 0x0, symName: _splash_screen, symObjAddr: 0x648, symBinAddr: 0x100002A3C, symSize: 0x24C }
-  - { offsetInCU: 0x484, offset: 0x484, size: 0x8, addend: 0x0, symName: _show_menu, symObjAddr: 0x894, symBinAddr: 0x100002C88, symSize: 0x18C }
-  - { offsetInCU: 0x4B6, offset: 0x4B6, size: 0x8, addend: 0x0, symName: _playAduio, symObjAddr: 0xA20, symBinAddr: 0x100002E14, symSize: 0x94 }
-  - { offsetInCU: 0x4EC, offset: 0x4EC, size: 0x8, addend: 0x0, symName: _main, symObjAddr: 0xAB4, symBinAddr: 0x100002EA8, symSize: 0x48 }
-  - { offsetInCU: 0x506, offset: 0x506, size: 0x8, addend: 0x0, symName: _GamePlay, symObjAddr: 0xAFC, symBinAddr: 0x100002EF0, symSize: 0xBF0 }
-  - { offsetInCU: 0x5A3, offset: 0x5A3, size: 0x8, addend: 0x0, symName: _ScoreBoard, symObjAddr: 0x16EC, symBinAddr: 0x100003AE0, symSize: 0x1A8 }
-...
